chore: recreate font elements with Typography
